LS-5332: Rename transport_httpjson to transport_httpthrift
