+++ src/imp/platform/browser/transport_httpthrift.js
